+++ src/components/Contents.tsx
@@ -1,7 +1,0 @@
-import { Box } from "@chakra-ui/react";
-
-function Contents() {
-  return <Box>Contents</Box>;
-}
-
-export default Contents;