+++ src/components/PostDetail.tsx
@@ -1,0 +1,19 @@
+import { Box, Flex } from "@chakra-ui/react";
+
+function PostDetail() {
+  return (
+    <Flex
+      flexDir="column"
+      align="center"
+      m="20px"
+      p="20px"
+      bg="whiteAlpha.600"
+      flexGrow="1"
+      borderRadius="10px"
+    >
+      Post Detail
+    </Flex>
+  );
+}
+
+export default PostDetail;