+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### v0.1.1-dev
 
+- [NEW] PostList and Post components / layout @christianmotta
 - [NEW] Creating a responsive layout using ChakraUI @christianmotta
 - [NEW] Setting up basic structure and libraries @christianmotta
 - [CHG] Project cleaned up @christianmotta